Chicago Budget Adventure
Inspiration
Budgeting can seem intimidating or boring. We wanted to show young people what budgeting might look like in a fun context. We chose to create a scenario where the user imagines they just got a job in Chicago as a young professional. They can practice making spending choices and see how those impact their ability to save towards their goal. We used Chicago as inspiration for all of the scenarios: the cost of living, transportation options, weekend outings are all real and accurate! Even the apartment choices are based on real available Chicago apartments we found on the website Zillow.

Unfortunately, financial literacy is not really taught in all schools. Young adults often never really learn how to budget or save money effectively. Especially for kids and teens in low-income neighborhoods in Chicago, access to financial literacy education is limited so we wanted to provide a fun way for them to get curious about it and want to learn more.

What it does
The website combines a "Choose Your Own Adventure" type story inspired by "Buzzfeed" type articles with a versatile budget tracker tool. It lets the user "play" at budgeting while they also learn more about the city of Chicago itself, typical costs of housing and food, the choices a young professional living on their own will have to make, and how to use banking accounts to grow their savings.

How we built it
We built the website using a combination of HTML, CSS, and Javascript languages as well as the Balsamiq, GitHub, Replit, and VSCode software.

Challenges we ran into
The time limit was definitely one of the biggest challenges we faced as there were many features we wanted to implement but unfortunately did not have enough time to do so. We were unable to implement the Javascript code features for the website where it would have tracked the user's choices throughout the adventure and calculated how much of their income they would have spent after picking their options and how much they would be able to save each month.

Accomplishments that we're proud of
The budget tracker calculator app where users can enter their income and expenses into the tracker and then see their bank balance. Having finished the basic layout for our choose your own savings adventure story. Using a brainwriting exercise to ideate in a way that included everyone and allowed us to collaborate.

What we learned
We learned how to plan site layouts beforehand on the Balsamiq software, how to use Replit to collaborate with team members, and how to use GitHub to publish our project.

What's next for Chicago Budget Adventure
Work on the JavaScript code to make the amount of money spent and saved throughout the adventure be trackable. Work on making the UI more accessible across multiple devices and also to make it easier to navigate.